requirements.txt
django
djangorestframework
django-jazzmin
djoser
drf_spectacular
psycopg2
psycopg2-binary